+++ PathTracingCUDA/include/Interval.h
@@ -6,20 +6,22 @@
 class Interval
 {
 public:
-	double min, max;
+	float min, max;
 
 	__device__ Interval();
-	__device__ Interval(double min, double max);
+	__device__ Interval(float min, float max);
 	__device__ Interval(const Interval& a, const Interval& b);
 
-	__device__ double Size() const;
-	__device__ bool Contains(double x) const;
+	__device__ float Size() const;
+	__device__ bool Contains(float x) const;
 
-	__device__ bool Surrounds(double x) const;
+	__device__ bool Surrounds(float x) const;
 
-	__device__ double Clamp(double x) const;
+	__device__ float Clamp(float x) const;
 
-	__device__ Interval Expand(double delta) const;
+	__device__ Interval Expand(float delta) const;
 
 	static const Interval Empty, Universe;
 };
+
+
