+++ PathTracingCUDA/include/Ray.h
@@ -7,17 +7,17 @@ class Ray
 public:
 	__device__ Ray() = default;
 
-	__device__ Ray(const glm::dvec3& origin, const glm::dvec3 direction) : orig(origin), dir(direction) {};
+	__device__ Ray(const glm::vec3& origin, const glm::vec3 direction) : orig(origin), dir(direction) {};
 
-	__device__ const glm::dvec3& origin() const { return this->orig; }
-	__device__ const glm::dvec3& direction() const { return this->dir; }
+	__device__ const glm::vec3& origin() const { return this->orig; }
+	__device__ const glm::vec3& direction() const { return this->dir; }
 
-	__device__ glm::dvec3 at(double t) const
+	__device__ glm::vec3 at(float t) const
 	{
 		return orig + t * dir;
 	}
 
 private:
-	glm::dvec3 orig;
-	glm::dvec3 dir;
+	glm::vec3 orig;
+	glm::vec3 dir;
 };
